Add easy-ecryptfs, an ecryptfs partition manager.
The default Ubuntu tool ecryptfs-setup-private is opinionated and
inflexible, as it's unable to create ecryptfs partitions except for
`~/.Private` => `~/Private`. This is strange, because the underlying
mount.ecryptfs_private tooling accepts an ALIAS argument and can handle
arbitrary directories.

easy-ecryptfs is a thin wrapper around those tools, making it much
easier to use ecryptfs in the same style as encfs. It has subcommands to
create, mount, and unmount ecryptfs filesystems.

Unlike ecryptfs-setup-private, which keeps critical key material in
~/.ecryptfs and not in the source directory, easy-ecryptfs saves all
important data in the source directory so that it can be put under
version control as a single repository for backup/replication if the
user so chooses.